Migrate delegated shell tests to Playwright
