+++ run-tests.js
@@ -43,10 +43,13 @@ const TEST_FILES = [
   // to Vitest (batch 22). The section-3b functional safeMarkerId-guard probes
   // (need live DOM + populated state) live in test-audit-dom.js below.
   'tests/test-audit-dom.js',
-  'tests/test-nav-delegated-actions-dom.js',
+  // test-nav-delegated-actions-dom.js moved to Playwright in
+  // tests/playwright/nav-delegated-actions.spec.js.
   'tests/test-dashboard-widget-delegated-actions-dom.js',
-  'tests/test-lens-page-shell-delegated-actions-dom.js',
-  'tests/test-light-page-view-delegated-actions-dom.js',
+  // test-lens-page-shell-delegated-actions-dom.js moved to Playwright in
+  // tests/playwright/lens-page-shell.spec.js.
+  // test-light-page-view-delegated-actions-dom.js moved to Playwright in
+  // tests/playwright/light-page-view.spec.js.
   // test-image-utils.js source-inspection + module-export checks ported to
   // Vitest (PR for batch 19). DOM-runtime assertions (sections 6 + 7)
   // moved to Playwright in tests/playwright/image-utils-dom.spec.js.
